# py/projects/user_api/openapi.json
{"openapi":"3.0.2","info":{"title":"Kodawari User API","description":"Provides CRUD routes for managing UserSchema objects.","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","description":"Retrieves the health status of the API.\n\nReturns:\n    A string of \"healthy\" if the API is healthy.","operationId":"get_health","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Health","type":"string"}}}}}}},"/user/{id}":{"get":{"summary":"Get","description":"Retrieves a user.\n\nArgs:\n    id: The identifier for the requested User resource.\nReturns:\n    A UserSchema, containing the requested User data.\nRaises:\n    HTTPException: The requested User resource could not be retrieved.","operationId":"get_user","parameters":[{"required":true,"schema":{"title":"Id","type":"integer"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user":{"post":{"summary":"Post","description":"Creates a user.\n\nPatches a user and sets the Location header to the created resource path, using properties defined in the user_create_request, a generated id, and a unix timestamp.\n\nArgs:\n    response: The FastAPI response class used for setting the Location header of the created resource path.\n    user_create_request: The UserSchema properties to set on the created resource.","operationId":"create_user","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete","description":"Deletes a user.\n\nDeletes a User resource using the ID retrieved from bearer_claims.\n\nArgs:\n    bearer_claims: The claims included on the Bearer token in the request.","operationId":"delete_user","responses":{"204":{"description":"Successful Response"}},"security":[{"HTTPBearer":[]}]},"patch":{"summary":"Patch","description":"Patches a user.\n\nPatches a user and sets the Location header to the modified resource path, using the ID retrieved from bearer_claims and properties defined in the user_patch_request.\n\nArgs:\n    response: The FastAPI response class used for setting the Location header of the modified resource path.\n    user_patch_request: The User resoruce properties to update.\n    bearer_claims: The claims included on the Bearer token in the request.","operationId":"patch_user","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPatchRequest"}}},"required":true},"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"UserCreateRequest":{"title":"UserCreateRequest","required":["display_name"],"type":"object","properties":{"display_name":{"title":"Display Name","maxLength":12,"minLength":4,"type":"string","description":"The non-unique alphanumeric display name of a user, restricted to a minimum of four and a maximum of twelve characters"},"description":{"title":"Description","maxLength":1000,"type":"string","description":"The user provided description for their profile, restricted to a maximum of 1000 characters, defaulting to empty string","default":""}},"description":"The required request body for creating a UserSchema object\n\nAttributes:\n    display_name: The non-unique alphanumeric display name of a user, restricted to a minimum of 4 and a maximum of 12 characters\n    description: The user provided description for their profile, restricted to a maximum of 1000 characters, defaulting to empty string"},"UserPatchRequest":{"title":"UserPatchRequest","type":"object","properties":{"display_name":{"title":"Display Name","maxLength":12,"minLength":4,"type":"string","description":"The non-unique alphanumeric display name of a user, restricted to a minimum of four and a maximum of twelve characters"},"description":{"title":"Description","maxLength":1000,"type":"string","description":"The user provided description for their profile, restricted to a maximum of 1000 characters, defaulting to empty string"}},"description":"The required request body for patching a UserSchema object\n\nAttributes:\n    display_name: The (optional) non-unique alphanumeric display name of a user, restricted to a minimum of 4 and a maximum of 12 characters\n    description: The (optional) user provided description for their profile, restricted to a maximum of 1000 characters, defaulting to empty string"},"UserSchema":{"title":"UserSchema","required":["id","display_name","description","joined"],"type":"object","properties":{"id":{"title":"Id","exclusiveMinimum":0.0,"type":"integer","description":"The unique identifier for a user"},"display_name":{"title":"Display Name","maxLength":12,"minLength":4,"type":"string","description":"The non-unique alphanumeric display name of a user, restricted to a minimum of four and a maximum of twelve characters"},"description":{"title":"Description","maxLength":1000,"type":"string","description":"The user provided description for their profile, restricted to a maximum of 1000 characters, defaulting to empty string"},"joined":{"title":"Joined","exclusiveMaximum":9999999999999.0,"exclusiveMinimum":1674484829053.0,"type":"integer","description":"The date a user was created, expressed as a unix timestamp in milliseconds"}},"description":"A user's profile information.\n\nAttributes:\n    id: The unique identifier for a user\n    display_name: The non-unique alphanumeric display name of a user, restricted to a minimum of 4 and a maximum of 12 characters\n    description: The user provided description for their profile, restricted to a maximum of 1000 characters, defaulting to empty string\n    joined: The date a user was created, expressed as a unix timestamp in milliseconds"},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}